renames files and event per pr suggestions
