+++ src/sidebar/modLauncher/ActiveSidebarModsListItem.module.scss
